requirements.txt
line-bot-sdk
flask
requests
beautifulsoup4
pyChatGPT
psycopg2
openai
imdbpy
